Updated Data to use Mabi
Moved some enums from Data to Mabi, that are Mabi specific.
