+++ src/ChannelServer/World/RegionCollision.cs
@@ -8,6 +8,7 @@ using System.Linq;
 using Aura.Channel.Util;
 using Aura.Data.Database;
 using Aura.Shared.Util;
+using Aura.Mabi.Const;
 
 namespace Aura.Channel.World
 {
