+++ src/Data/Database/Items.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json.Linq;
+using Aura.Mabi.Const;
 
 namespace Aura.Data.Database
 {
@@ -51,7 +52,7 @@ namespace Aura.Data.Database
 		public byte InjuryMax { get; set; }
 		public sbyte Critical { get; set; }
 		public byte Balance { get; set; }
-		public byte AttackSpeed { get; set; }
+		public AttackSpeed AttackSpeed { get; set; }
 		public byte KnockCount { get; set; }
 
 		public int BagWidth { get; set; }
@@ -140,7 +141,7 @@ namespace Aura.Data.Database
 				info.InjuryMax = entry.ReadByte("injuryMax");
 				info.Critical = entry.ReadSByte("critical");
 				info.Balance = entry.ReadByte("balance");
-				info.AttackSpeed = entry.ReadByte("attackSpeed");
+				info.AttackSpeed = (AttackSpeed)entry.ReadByte("attackSpeed");
 				info.KnockCount = entry.ReadByte("knockCount");
 			}
 
@@ -172,110 +173,5 @@ namespace Aura.Data.Database
 
 			this.Entries[info.Id] = info;
 		}
-	}
-
-	public enum ItemType
-	{
-		Armor = 0,
-		Headgear = 1,
-		Glove = 2,
-		Shoe = 3,
-		Book = 4,
-		Currency = 5,
-		ItemBag = 6,
-		Weapon = 7,
-		Weapon2H = 8, // 2H, bows, tools, etc
-		Weapon2 = 9, // Ineffective Weapons? Signs, etc.
-		Instrument = 10,
-		Shield = 11,
-		Robe = 12,
-		Accessory = 13,
-		SecondaryWeapon = 14,
-		MusicScroll = 15,
-		Manual = 16,
-		EnchantScroll = 17,
-		CollectionBook = 18,
-		ShopLicense = 19,
-		FaliasTreasure = 20,
-		Kiosk = 21,
-		StyleArmor = 22,
-		StyleHeadgear = 23,
-		StyleGlove = 24,
-		StyleShoe = 25,
-		ComboCard = 27,
-		Unknown2 = 28,
-		Hair = 100,
-		Face = 101,
-		Usable = 501,
-		Quest = 502,
-		Usable2 = 503,
-		Unknown1 = 504,
-		Sac = 1000,
-		Misc = 1001,
-	}
-
-	public enum StackType
-	{
-		None = 0,
-		Stackable = 1,
-		Sac = 2,
-	}
-
-	public enum InstrumentType
-	{
-		Lute = 0,
-		Ukulele = 1,
-		Mandolin = 2,
-		Whistle = 3,
-		Roncadora = 4,
-		Flute = 5,
-		Chalumeau = 6,
-
-		ToneBottleC = 7,
-		ToneBottleD = 8,
-		ToneBottleE = 9,
-		ToneBottleF = 10,
-		ToneBottleG = 11,
-		ToneBottleB = 12,
-		ToneBottleA = 13,
-
-		Tuba = 18,
-		Lyra = 19,
-		ElectricGuitar = 20,
-
-		Piano = 21,
-		Violin = 22,
-		Cello = 23,
-
-		BassDrum = 66,
-		Drum = 67,
-		Cymbals = 68,
-
-		HandbellC = 69,
-		HandbellD = 70,
-		HandbellE = 71,
-		HandbellF = 72,
-		HandbellG = 73,
-		HandbellB = 74,
-		HandbellA = 75,
-		HandbellHighC = 76,
-
-		Xylophone = 77,
-
-		MaleVoiceKr1 = 81,
-		MaleVoiceKr2 = 82,
-		MaleVoiceKr3 = 83,
-		MaleVoiceKr4 = 84,
-		FemaleVoiceKr1 = 90,
-		FemaleVoiceKr2 = 91,
-		FemaleVoiceKr3 = 92,
-		FemaleVoiceKr4 = 93,
-		FemaleVoiceKr5 = 94,
-
-		MaleChorusVoice = 100,
-		FemaleChorusVoice = 110,
-
-		MaleVoiceJp = 120,
-		FemaleVoiceJp = 121,
 	}
 }
