+++ src/Mabi/Const/Races.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) Aura development team - Licensed under GNU GPL
+// For more information, see license file in the main folder
+
+namespace Aura.Mabi.Const
+{
+	/// <summary>
+	/// Race's gender
+	/// </summary>
+	public enum Gender : byte
+	{
+		None,
+		Female,
+		Male,
+		Universal,
+	}
+}
